Fix/CVE 2026 34986 go jose go OIDC (#581)
* fix(deps): address CVE-2026-34986 and drop go-jose/v3

- Bump github.com/go-jose/go-jose/v4 to v4.1.4 (patched for CVE-2026-34986).
- Upgrade github.com/coreos/go-oidc/v3 to v3.17.0 so the OIDC stack uses
  go-jose/v4 only; removes the indirect go-jose/v3 dependency.

Made-with: Cursor

* chore: update aws

* chore: fix tests

+++ internal/config/config.go
@@ -254,7 +254,7 @@ type Config struct {
 
 	// Rate Limiting
 	// RateLimitRPS is the maximum requests per second per IP
-	RateLimitRPS float64
+	RateLimitRPS int
 	// RateLimitBurst is the maximum burst size per IP
 	RateLimitBurst int
 	// RateLimitFailClosed rejects requests when the rate limit backend errors (default: fail-open).
